predicting-mobile-phones-price-range/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
Pillow
numpy
scikit-learn==1.2.2